+++ srcs/utils/unit.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   unit.c                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sharnvon <sharnvon@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/03/10 00:13:49 by sharnvon          #+#    #+#             */
+/*   Updated: 2023/03/10 00:13:50 by sharnvon         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minirt.h"
 
 t_color	set_color(int r, int g, int b)
@@ -10,7 +22,7 @@ t_color	set_color(int r, int g, int b)
 	return (color);
 }
 
-t_size set_size(double w, double h, double d)
+t_size	set_size(float w, float h, float d)
 {
 	t_size	size;
 
@@ -30,7 +42,7 @@ t_ray	set_ray(t_vtr a, t_vtr b)
 	return (r);
 }
 
-int		close0(const float f1, const float f2)
+int	close0(const float f1, const float f2)
 {
 	if (fabsf(f1 - f2) < FLT_EPSILON)
 		return (1);
